add verification is the task is stopped in ECS, just proceed with updating the service

+++ ops_boba/cfn-devenv.sh
@@ -486,18 +486,26 @@ function destroy_dev_services {
             SERVICE4RESTART=`aws ecs list-services --region ${REGION} --cluster $ECS_CLUSTER|grep -i ${ENV_PREFIX}|cut -d/ -f3|sed 's#,##g'|egrep -vi ^datadog|tr '\n' ' '|sed 's#"##g'`
               if [[ "${SERVICE_NAME}" == "data-transport-layer" || "${SERVICE_NAME}" == "replica-dtl" || "${SERVICE_NAME}" == "verifier-dtl" || "${SERVICE_NAME}" == "replica-dtl-bkp01" || "${SERVICE_NAME}" == "replica-dtl-bkp02"  ]]; then
                 TASK_ARN=`aws ecs list-tasks --cluster $ECS_CLUSTER --service ${SERVICE_NAME} --output text --query taskArns[0]`
-                aws ecs update-service  --region ${REGION} --service ${SERVICE_NAME} --cluster $ECS_CLUSTER --desired-count 0 >> /dev/null
-                aws ecs stop-task --cluster $ECS_CLUSTER --task $TASK_ARN >> /dev/null
-                aws ssm send-command --document-name "AWS-RunShellScript" --instance-ids $EC2_INSTANCE --parameters commands="rm -rf /mnt/efs/db/LOCK" --region ${REGION} --output text >> /dev/null
-                aws ssm send-command --document-name "AWS-RunShellScript" --instance-ids $EC2_INSTANCE --parameters commands="rm -rf /mnt/efs/replica-dtl/db/LOCK" --region ${REGION} --output text >> /dev/null
-                aws ssm send-command --document-name "AWS-RunShellScript" --instance-ids $EC2_INSTANCE --parameters commands="rm -rf /mnt/efs/verifier-dtl/db/LOCK" --region ${REGION} --output text >> /dev/null
+                if [[ $TASK_ARN == "None" ]]; then
+                  continue
+                else
+                  aws ecs update-service  --region ${REGION} --service ${SERVICE_NAME} --cluster $ECS_CLUSTER --desired-count 0 >> /dev/null
+                  aws ecs stop-task --cluster $ECS_CLUSTER --task $TASK_ARN >> /dev/null
+                  aws ssm send-command --document-name "AWS-RunShellScript" --instance-ids $EC2_INSTANCE --parameters commands="rm -rf /mnt/efs/db/LOCK" --region ${REGION} --output text >> /dev/null
+                  aws ssm send-command --document-name "AWS-RunShellScript" --instance-ids $EC2_INSTANCE --parameters commands="rm -rf /mnt/efs/replica-dtl/db/LOCK" --region ${REGION} --output text >> /dev/null
+                  aws ssm send-command --document-name "AWS-RunShellScript" --instance-ids $EC2_INSTANCE --parameters commands="rm -rf /mnt/efs/verifier-dtl/db/LOCK" --region ${REGION} --output text >> /dev/null
+                fi
               elif [[ "${SERVICE_NAME}" == "l2geth" || "${SERVICE_NAME}" == "replica-l2" || "${SERVICE_NAME}" == "verifier-l2" || "${SERVICE_NAME}" == "replica-l2-bkp01" || "${SERVICE_NAME}" == "replica-l2-bkp02" ]]; then
                 TASK_ARN=`aws ecs list-tasks --cluster $ECS_CLUSTER --service ${SERVICE_NAME} --output text --query taskArns[0]`
-                aws ecs update-service  --region ${REGION} --service ${SERVICE_NAME} --cluster $ECS_CLUSTER --desired-count 0 >> /dev/null
-                aws ecs stop-task --cluster $ECS_CLUSTER --task $TASK_ARN >> /dev/null
-                aws ssm send-command --document-name "AWS-RunShellScript" --instance-ids $EC2_INSTANCE --parameters commands="rm -rf /mnt/efs/geth_l2/geth/LOCK" --region ${REGION} --output text >> /dev/null
-                aws ssm send-command --document-name "AWS-RunShellScript" --instance-ids $EC2_INSTANCE --parameters commands="rm -rf /mnt/efs/replica-l2/geth/LOCK" --region ${REGION} --output text >> /dev/null
-                aws ssm send-command --document-name "AWS-RunShellScript" --instance-ids $EC2_INSTANCE --parameters commands="rm -rf /mnt/efs/verifier-l2/geth/LOCK" --region ${REGION} --output text >> /dev/null
+                if [[ $TASK_ARN == "None" ]]; then
+                  continue
+                else
+                  aws ecs update-service  --region ${REGION} --service ${SERVICE_NAME} --cluster $ECS_CLUSTER --desired-count 0 >> /dev/null
+                  aws ecs stop-task --cluster $ECS_CLUSTER --task $TASK_ARN >> /dev/null
+                  aws ssm send-command --document-name "AWS-RunShellScript" --instance-ids $EC2_INSTANCE --parameters commands="rm -rf /mnt/efs/geth_l2/geth/LOCK" --region ${REGION} --output text >> /dev/null
+                  aws ssm send-command --document-name "AWS-RunShellScript" --instance-ids $EC2_INSTANCE --parameters commands="rm -rf /mnt/efs/replica-l2/geth/LOCK" --region ${REGION} --output text >> /dev/null
+                  aws ssm send-command --document-name "AWS-RunShellScript" --instance-ids $EC2_INSTANCE --parameters commands="rm -rf /mnt/efs/verifier-l2/geth/LOCK" --region ${REGION} --output text >> /dev/null
+                fi
               else
                 aws ecs update-service  --region ${REGION} --service ${SERVICE_NAME} --cluster $ECS_CLUSTER --desired-count 0 >> /dev/null
               fi
